refactor: split admin scheduling workspace api/view types for line budget
